refactor(runtime): split kernel command dispatch into router + callback bundles
Extract the in-line request_handler cached_property and the six *Callbacks
classes out of runtime.py (~1270 LOC removed) into a small RequestRouter,
a KernelRequestHandlers bundle for handlers that need kernel-side context
wrappers, and per-domain callback modules under _runtime/callbacks/ that
satisfy a KernelCallback protocol and self-register their command bindings.
