+++ marimo/_runtime/callbacks/__init__.py
@@ -1,0 +1,22 @@
+# Copyright 2026 Marimo. All rights reserved.
+from __future__ import annotations
+
+from marimo._runtime.callbacks.cache import CacheCallbacks
+from marimo._runtime.callbacks.datasets import DatasetCallbacks
+from marimo._runtime.callbacks.external_storage import (
+    ExternalStorageCallbacks,
+)
+from marimo._runtime.callbacks.packages import PackagesCallbacks
+from marimo._runtime.callbacks.protocol import KernelCallback
+from marimo._runtime.callbacks.secrets import SecretsCallbacks
+from marimo._runtime.callbacks.sql import SqlCallbacks
+
+__all__ = [
+    "CacheCallbacks",
+    "DatasetCallbacks",
+    "ExternalStorageCallbacks",
+    "KernelCallback",
+    "PackagesCallbacks",
+    "SecretsCallbacks",
+    "SqlCallbacks",
+]
